UT comms: helper function moved into committees file, utils.py deleted
